% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{close_to}
\alias{close_to}
\title{Weight: how close to you to a certain point}
\usage{
close_to(
  board,
  i_star = ceiling(dim(board)[1]/2),
  j_star = ceiling(dim(board)[1]/2)
)
}
\description{
Weight: how close to you to a certain point
}
