% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{density_local}
\alias{density_local}
\title{Weight: are there moles in a star-pattern around you?}
\usage{
density_local(board, i, j, degree = 1)
}
\description{
Weight: are there moles in a star-pattern around you?
}
