% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{linear_left}
\alias{linear_left}
\title{Weight: how far are you from the left-most column?}
\usage{
linear_left(board, n = dim(board)[1])
}
\description{
Weight: how far are you from the left-most column?
}
