% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{print.board}
\alias{print.board}
\title{Print method for 'board' class}
\usage{
\method{print}{board}(board, ...)
}
\arguments{
\item{board}{An object of class 'board'.}

\item{...}{Additional arguments (can ignore for now)}
}
\value{
The input object 'board' (invisibly)
}
\description{
Implements the custom print behavior for whack-a-mole boards
}
