% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radial_top_left}
\alias{radial_top_left}
\title{Weights relative to the position of a cell in a board}
\usage{
radial_top_left(board, n = dim(board)[1])
}
\description{
Weights relative to the position of a cell in a board
}
