% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_edges}
\alias{strip_edges}
\title{Remove edges around a board}
\usage{
strip_edges(board)
}
\description{
remove the edges of a board and return it's sub-board
(useful for turning a board back to its original after making it big)
}
