% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{density_local}
\alias{density_local}
\title{The density of moles across a matrix, with strong local preference}
\usage{
density_local(board, i, j, degree = 1)
}
\description{
The density of moles across a matrix, with strong local preference
}
