% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play a game of Whack-A-Mole}
\usage{
play(n = 3, theta = 0.8)
}
\description{
Initiates a random board of moles, then iterates over an algorithm to whack moles,
with the goal of whacking all moles within a certain number of whacks
}
