% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{radial_top_left}
\alias{radial_top_left}
\title{Weight: how far you are from the top-left corner?}
\usage{
radial_top_left(board, n = dim(board)[1])
}
\description{
Weight: how far you are from the top-left corner?
}
